==> rcodes/0. project setup.R <==
## run for the first time, and require data from Wenli, Yong or Yingbing
## for setup folders


dir.create(file.path(".", "rcodes"))
dir.create(file.path(".", "data"))
dir.create(file.path(".", "data", "rawdata"))
dir.create(file.path(".", "data", "compileddata"))
dir.create(file.path(".", "results"))
dir.create(file.path(".", "documents"))
dir.create(file.path(".", "reference"))
